Fixed a compiler warning Increased Config doc size for ESP32 builds

+++ src/output/OutputMgr.hpp
@@ -95,9 +95,9 @@ private:
 #   define OM_CHANNEL_TYPE_NAME    F("type")
 
 #ifdef ARDUINO_ARCH_ESP32
-#   define OM_MAX_CONFIG_SIZE      (4*1024)
+#   define OM_MAX_CONFIG_SIZE      ((size_t)(4*1024))
 #else
-#   define OM_MAX_CONFIG_SIZE      (3*1024)
+#   define OM_MAX_CONFIG_SIZE      ((size_t)(3*1024))
 #endif // def ARDUINO_ARCH_ESP32
 
     bool HasBeenInitialized = false;
